% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{astir_paper_theme}
\alias{astir_paper_theme}
\title{Create a theme for all plot for the astir paper}
\usage{
astir_paper_theme()
}
\description{
Create a theme for all plot for the astir paper
}
\examples{
astir_paper_theme()
}
