% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{createHeatmap}
\alias{createHeatmap}
\title{Create an expression heatmap}
\usage{
createHeatmap(
  sce,
  cell_type_column = "cell_type",
  assay = "logcounts",
  thresh = 2
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} containing expression values}

\item{cell_type_column}{The column of \code{colData(sce)} that contains 
cell type annotation}

\item{assay}{The assay of \code{sce} containg expression data to plot}

\item{thresh}{The threshold to winsorize expression values at}
}
\value{
A heatmap of class \code{ComplexHeatmap::Heatmap}
}
\description{
Note that this creates a \code{Heatmap} object. To plot,
call \code{draw(...)} on the returned object.
}
\examples{
        counts <- matrix(rpois(100, lambda = 10), ncol=10, nrow=10)
}
