% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cell_assignments}
\alias{get_cell_assignments}
\title{Given a celltype probability matrix, return most likely
cell types (with probability above threshold).}
\usage{
get_cell_assignments(mat, threshold = 0.7)
}
\arguments{
\item{mat}{Cell-by-celltype probability matrix}

\item{threshold}{Probability threshold above which a cell is assigned
to a celltype

 @export}
}
\description{
If no celltype has probability \code{> threshold} for a given
cell type, the cell is assigned to "Unknown"
}
