% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_celltypes}
\alias{get_celltypes}
\title{Get cell type assignments given probability matrix}
\usage{
get_celltypes(prob_mat, thresh = 0.7)
}
\arguments{
\item{prob_mat}{A probability matrix?}

\item{thresh}{A measure for cell type assignment?}
}
\description{
Get cell type assignments given probability matrix
}
\examples{
counts <- matrix(rpois(100, lambda = 10), ncol=10, nrow=10)
}
