% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{imc_exclusive_marker_score}
\alias{imc_exclusive_marker_score}
\title{Get a QC score for the mutually exclusive expression of 2 markers}
\usage{
imc_exclusive_marker_score(
  sce,
  marker_1 = "pan Cytokeratin",
  marker_2 = "CD45",
  assay_name = "logcounts"
)
}
\arguments{
\item{sce}{An input \code{SummarizedExperiment} or \code{SingleCellExperiment}}

\item{marker_1}{The name of the first marker (normally keratin)}

\item{marker_2}{The name of the second marker (normally CD45/PTPRC)}
}
\description{
Implements a very basic QC test: CD45 and keratin should exhibit mutually
exclusive expression, so performs a robust regression to check this trend,
and reports the z-statistic. The more negative this is, the better the
mutually-exclusive expression (to a first approximation).
}
