% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jackson_2020_markers}
\alias{jackson_2020_markers}
\title{Marker information from Jackson et al. (Nature 2020)}
\format{An object of class \code{list} of length 2.}
\usage{
jackson_2020_markers
}
\description{
Marker information from Jackson et al. (Nature 2020)
}
\examples{
data(jackson_2020_markers)
}
\keyword{datasets}
