% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jackson_2020_tag_target}
\alias{jackson_2020_tag_target}
\title{Map of tag to targets for Jackson et al. (Nature 2020)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 2 columns.}
\usage{
jackson_2020_tag_target
}
\description{
Map of tag to targets for Jackson et al. (Nature 2020)
}
\examples{
data(jackson_2020_tag_target)
}
\keyword{datasets}
