% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{patient_clinSubtype_colours}
\alias{patient_clinSubtype_colours}
\title{Color maps of patient clinical subtypes}
\usage{
patient_clinSubtype_colours()
}
\description{
Color maps of patient clinical subtypes
}
\examples{
patient_clinSubtype_colours()
}
