% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{patient_clinSubtype_colours_wagner}
\alias{patient_clinSubtype_colours_wagner}
\title{Set Wagner colour maps to patient clinical subtypes}
\usage{
patient_clinSubtype_colours_wagner()
}
\description{
Set Wagner colour maps to patient clinical subtypes
}
\examples{
patient_clinSubtype_colours_wagner()
}
