% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{patient_response_colours}
\alias{patient_response_colours}
\title{Colour maps of patient responses}
\usage{
patient_response_colours()
}
\description{
Colour maps of patient responses
}
\examples{
patient_response_colours()
}
