% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSpatial}
\alias{plotSpatial}
\title{Plot the spatial variation in gene expression or colData}
\usage{
plotSpatial(
  sce,
  colour_by = NULL,
  x_str = "Location_Center_X",
  y_str = "Location_Center_Y",
  winsorize = FALSE,
  w_limits = c(0.05, 0.95),
  exprs_values = "logcounts",
  ...
)
}
\arguments{
\item{sce}{SingleCellExperiment containing IMC data}

\item{colour_by}{Either a feature name or column of \code{colData(sce)}}

\item{x_str}{The column of \code{colData(sce)} that codes for the x coordinate}

\item{y_str}{The column of \code{colData(sce)} that codes for the Y coordinate}

\item{...}{Additional arguments to pass to \code{plotColData}}
}
\description{
Plot the spatial variation in gene expression or colData
}
