% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{print_marker_info}
\alias{print_marker_info}
\title{Print information about marker genes}
\usage{
print_marker_info(markers)
}
\arguments{
\item{markers}{A marker list from \code{read_markers}}
}
\value{
Prints marker information
}
\description{
Print cell type and cell state marker info
}
\examples{
        counts <- matrix(rpois(100, lambda = 10), ncol=10, nrow=10)
}
