% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{read_markers}
\alias{read_markers}
\title{Read a marker yaml file}
\usage{
read_markers(path)
}
\arguments{
\item{path}{Input file path (yaml file)}
}
\value{
A marker list including cell type and state markers
}
\description{
Read a marker yaml file
}
