% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_rownames_jackson}
\alias{tidy_rownames_jackson}
\title{Convert from original SC_Dat.csv rownames to understandable ones}
\usage{
tidy_rownames_jackson(sce)
}
\arguments{
\item{sce}{Input SingleCellExperiment}
}
\description{
Convert from original SC_Dat.csv rownames to understandable ones
}
