% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_csv}
\alias{to_csv}
\title{Write to a CSV file to be read in by python}
\usage{
to_csv(
  sce,
  file,
  include_xy = TRUE,
  x_str = "Location_Center_X",
  y_str = "Location_Center_Y",
  exprs_values = "logcounts"
)
}
\arguments{
\item{sce}{SingleCellExperiment containing IMC data}

\item{file}{The output file to write to}

\item{include_xy}{Logical: should spatial location be written as first two
columns?}

\item{x_str}{The column of \code{colData(sce)} that codes for the x coordinate}

\item{y_str}{The column of \code{colData(sce)} that codes for the Y coordinate}

\item{exprs_values}{The type of the SingleCellExperiment?}
}
\description{
Writes the cell names as row names, and the (X,Y) co-ordinates
as the first two columns (named \code{X} and \code{Y}).
}
