% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize the (log) counts}
\usage{
winsorize(sce, exprs_values = "logcounts", w_limits = c(0.05, 0.95))
}
\arguments{
\item{sce}{An input \code{SummarizedExperiment} or \code{SingleCellExperiment}}

\item{exprs_values}{Type of input?}

\item{w_limits}{cut off limits?}
}
\description{
Winsorize the (log) counts
}
\examples{
counts <- matrix(rpois(100, lambda = 10), ncol=10, nrow=10)
}
