% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{imc_exclusive_marker_score}
\alias{imc_exclusive_marker_score}
\title{Get a QC score for the mutually exclusive expression of 2 markers}
\usage{
imc_exclusive_marker_score(
  sce,
  marker_1 = "pan Cytokeratin",
  marker_2 = "CD45",
  assay_name = "logcounts"
)
}
\description{
Get a QC score for the mutually exclusive expression of 2 markers
}
